mac-requirements.txt
python-dotenv
# tensorflow>=2.8
numpy
pandas
matplotlib
Pillow
kagglehub
json5
tensorflow-metal
tensorflow-macos
tqdm
mlflow
fastapi
uvicorn
tenacity
python-multipart